Refactor TMCS and Owen into separate modules
