+++ src/pydvl/value/shapley/actor.py
@@ -17,7 +17,7 @@ from pydvl.utils.config import ParallelConfig
 from pydvl.utils.parallel.actor import Coordinator, RayActorWrapper, Worker
 from pydvl.utils.utility import Utility
 from pydvl.value.result import ValuationResult
-from pydvl.value.shapley.montecarlo import TruncationPolicy
+from pydvl.value.shapley.truncated import TruncationPolicy
 from pydvl.value.stopping import MaxChecks, StoppingCriterion
 
 __all__ = ["get_shapley_coordinator", "get_shapley_worker"]
